requirements.txt
accelerate==1.12.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.1.0
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
joblib==1.5.3
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
psutil==7.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
soupsieve==2.8.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
urllib3==2.6.3
